added version into the serialized sample

+++ src/buffer_loader/io.rs
@@ -11,6 +11,9 @@ use commons::io::write_all;
 use super::LockedBuffer;
 
 
+type VersionedSample = (usize, Vec<ExampleWithScore>);
+
+
 const FILENAME: &str = "sample.bin";
 const REGION:   &str = "us-west-1";
 const BUCKET:   &str = "tmsn-cache";
@@ -23,6 +26,7 @@ const S3_PATH:  &str = "samples/";
 pub fn write_memory(
     new_sample: Vec<ExampleWithScore>,
     new_sample_buffer: LockedBuffer,
+    _version: usize,
 ) {
     let new_sample_lock = new_sample_buffer.write();
     *(new_sample_lock.unwrap()) = Some(new_sample);
@@ -32,9 +36,11 @@ pub fn write_memory(
 pub fn write_local(
     new_sample: Vec<ExampleWithScore>,
     _new_sample_buffer: LockedBuffer,
+    version: usize,
 ) {
     let filename = FILENAME.to_string() + "_WRITING";
-    write_all(&filename, &serialize(&new_sample).unwrap())
+    let data: VersionedSample = (version, new_sample);
+    write_all(&filename, &serialize(&data).unwrap())
         .expect("Failed to write the sample set to file");
     rename(filename, FILENAME.to_string()).unwrap();
 }
@@ -43,6 +49,7 @@ pub fn write_local(
 pub fn write_s3(
     new_sample: Vec<ExampleWithScore>,
     _new_sample_buffer: LockedBuffer,
+    version: usize,
 ) {
     let region = REGION.parse().unwrap();
     // TODO: Add support to read credentials from the config file
@@ -53,8 +60,9 @@ pub fn write_s3(
 
     // In case the file exists, delete it
     let (_, _) = bucket.delete(&filename).unwrap();
+    let data: VersionedSample = (version, new_sample);
     let (_, code) = bucket.put(
-        &filename, &serialize(&new_sample).unwrap(), "application/octet-stream").unwrap();
+        &filename, &serialize(&data).unwrap(), "application/octet-stream").unwrap();
     assert_eq!(200, code);
 }
 
@@ -63,21 +71,28 @@ pub fn write_s3(
 
 pub fn load_local(
     new_sample_buffer: LockedBuffer,
-) {
+    last_version: usize,
+) -> usize {
     let ori_filename = FILENAME.to_string();
     let filename = ori_filename.clone() + "_READING";
     if rename(ori_filename, filename.clone()).is_ok() {
-        let new_sample: Vec<ExampleWithScore> = deserialize(read_all(&filename).as_ref()).unwrap();
-        let new_sample_lock = new_sample_buffer.write();
-        *(new_sample_lock.unwrap()) = Some(new_sample);
-        remove_file(filename).unwrap();
+        let (version, new_sample): VersionedSample =
+            deserialize(read_all(&filename).as_ref()).unwrap();
+        if version > last_version {
+            let new_sample_lock = new_sample_buffer.write();
+            *(new_sample_lock.unwrap()) = Some(new_sample);
+            remove_file(filename).unwrap();
+            return version;
+        }
     }
+    last_version
 }
 
 
 pub fn load_s3(
     new_sample_buffer: LockedBuffer,
-) {
+    last_version: usize,
+) -> usize {
     let region = REGION.parse().unwrap();
     // TODO: Add support to read credentials from the config file
     // Read credentials from the environment variables
@@ -89,6 +104,11 @@ pub fn load_s3(
     let (data, code) = bucket.get(&filename).unwrap();
     if code == 200 {
         let new_sample_lock = new_sample_buffer.write();
-        *(new_sample_lock.unwrap()) = Some(deserialize(&data).unwrap());
+        let (version, data) = deserialize(&data).unwrap();
+        if version > last_version {
+            *(new_sample_lock.unwrap()) = Some(data);
+            return version;
+        }
     }
+    last_version
 }
